[Fix] Fix openmpt123 stdout output with int16 samples.

+++ gcc-fixes/openmpt123/openmpt123.cpp
@@ -307,7 +307,7 @@ public:
 				interleaved_int_buffer.push_back( buffers[channel][frame] );
 			}
 		}
-		std::cout.write( (const char*)( interleaved_int_buffer.data() ), interleaved_int_buffer.size() * sizeof( std::int32_t ) );
+		std::cout.write( (const char*)( interleaved_int_buffer.data() ), interleaved_int_buffer.size() * sizeof( std::int16_t ) );
 	}
 };
